refactor(ext/node): consolidate _fs_watch, _fs_utimes, _fs_truncate, _fs_symlink into fs.ts
Follow-up to #32588. Inlines the contents of _fs/_fs_watch.ts,
_fs/_fs_utimes.ts, _fs/_fs_truncate.ts, and _fs/_fs_symlink.ts into
the main fs.ts polyfill, reducing file fragmentation. Updates
promises.ts and cp_sync.ts to import from node:fs instead.
