+++ ext/node/lib.rs
@@ -394,10 +394,6 @@ deno_core::extension!(deno_node,
     "_fs/_fs_realpath.ts",
     "_fs/_fs_stat.ts",
     "_fs/_fs_statfs.ts",
-    "_fs/_fs_symlink.ts",
-    "_fs/_fs_truncate.ts",
-    "_fs/_fs_utimes.ts",
-    "_fs/_fs_watch.ts",
     "_next_tick.ts",
     "_process/exiting.ts",
     "_process/process.ts",
