+++ ext/node/polyfills/_fs/cp/cp_sync.ts
@@ -6,8 +6,7 @@ import { copyFileSync } from "ext:deno_node/_fs/_fs_copy.ts";
 import { existsSync } from "ext:deno_node/_fs/_fs_exists.ts";
 import { mkdirSync, opendirSync } from "node:fs";
 import { readlinkSync } from "ext:deno_node/_fs/_fs_readlink.ts";
-import { symlinkSync } from "ext:deno_node/_fs/_fs_symlink.ts";
-import { utimesSync } from "ext:deno_node/_fs/_fs_utimes.ts";
+import { symlinkSync, utimesSync } from "node:fs";
 import {
   ERR_FS_CP_DIR_TO_NON_DIR,
   ERR_FS_CP_EEXIST,
