+++ tools/core_import_map.json
@@ -545,9 +545,6 @@
     "ext:deno_node/_fs/_fs_readv.ts": "../ext/node/polyfills/_fs/_fs_readv.ts",
     "ext:deno_node/_fs/_fs_stat.ts": "../ext/node/polyfills/_fs/_fs_stat.ts",
     "ext:deno_node/_fs/_fs_statfs.ts": "../ext/node/polyfills/_fs/_fs_statfs.ts",
-    "ext:deno_node/_fs/_fs_symlink.ts": "../ext/node/polyfills/_fs/_fs_symlink.ts",
-    "ext:deno_node/_fs/_fs_utimes.ts": "../ext/node/polyfills/_fs/_fs_utimes.ts",
-    "ext:deno_node/_fs/_fs_watch.ts": "../ext/node/polyfills/_fs/_fs_watch.ts",
     "ext:deno_node/_global.d.ts": "../ext/node/polyfills/_global.d.ts",
     "ext:deno_node/_next_tick.ts": "../ext/node/polyfills/_next_tick.ts",
     "ext:deno_node/_process/exiting.ts": "../ext/node/polyfills/_process/exiting.ts",
